Split parser into smaller files for faster compilation
